Complete tests of EBSD refinement with NLopt
