+++ kikuchipy/__init__.py
@@ -21,11 +21,13 @@ from kikuchipy.release import version as __version__
 # Attempt import only once
 try:
     import pyvista
+
     _pyvista_installed = True
 except ImportError:  # pragma: no cover
     _pyvista_installed = False
 try:
     import nlopt
+
     _nlopt_installed = True
 except ImportError:  # pragma: no cover
     _nlopt_installed = False
