--- kikuchipy/indexing/tests/test_refinement.py
@@ -1,30 +1,0 @@
-# Copyright 2019-2022 The kikuchipy developers
-#
-# This file is part of kikuchipy.
-#
-# kikuchipy is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# kikuchipy is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
-
-import numpy as np
-from orix.crystal_map import Phase
-import pytest
-
-import kikuchipy as kp
-
-
-class TestEBSDRefinementSetup:
-    pass
-
-
-class TestRefinementSolverNumba:
-    pass
