+++ kikuchipy/pattern/tests/test_pattern.py
@@ -32,6 +32,7 @@ from kikuchipy.pattern._pattern import (
     remove_dynamic_background,
     _dynamic_background_frequency_space_setup,
     _get_image_quality_numba,
+    _mask_pattern,
     _remove_background_subtract,
     _remove_background_divide,
     _remove_static_background_subtract,
@@ -511,3 +512,15 @@ class TestNormalizeIntensityPattern:
 
         assert np.allclose(np.mean(p2), 0, atol=1e-6)
         assert np.allclose(p2, answer, atol=1e-4)
+
+
+class TestMaskPattern:
+    def test_mask_pattern_numba(self):
+        p = np.arange(100).astype("float32")
+        mask = np.ones(p.size, dtype=bool)
+        mask[:10] = False
+        p_masked = _mask_pattern(p, mask)
+        p_masked2 = _mask_pattern.py_func(p, mask)
+        assert p_masked.size == 90
+        assert np.isclose(p_masked.mean(), 54.5)
+        assert np.allclose(p_masked, p_masked2)
